Physics update, nr 7
Made GameObjects easier to instanciate with colliders and rigidbodies, so it can be done within 2 lines of code, still room for improvement. Added dynamic programming for the collision checking, from 40k calls to 16k calls with 4 GameObjects, scales even more with higher number of GameObjects. Added that you can remove components from GameObjects, at least from their list of components. They still exist in memory as they can have living references to the GameObject and the other components on it, if not also other GameObjects and their components. Don't know how to solve this one. Can't simply remove a collider and expect the GameObject to have gotten removed the collision functionality, it's just removed from a simple list, but it still exist