Started on scene window
Made a scene class that can be added GameObjects for it to hold and manage.
Made a fps counter for debuging purpose and perfomance overview, tried reading and writing from a file, wich was successfull, a problem tough is that pointers can't be used as data when saved as the pointer will be changed when reloading the scene. Need some kind of instantiation of objects from file by copying there data and references, not a simple task, tough object position saving could be quite easily done, which would work fine as well. Need to make the scene window usable to manipulate position and creation of objects