Physics thread
Made physics run in it's own thread, which should be called at a fixed time rate, for now it does not, it is jittery. It does work, but is not smooth in it's movement and strictly speaking unaceptable condition for real game development. Did also add the particle system to the physics update, though the particle systems should potentionally also themselfs be running in their own threads, but still be handled by the physics system. Also made particle system a GameObject so it can easily be added to the physics system and updated without interference from a developer