Added macros for debugging and singletons, Particle updates
Added macros to debugging so when you run in debug you can use uniquely named keywords you yourself make for easy to read code. In release build all debugging will be removed automaticly by the simple use of macros.
Made some particle system updates, made it use universal time and whatnot. The async could be used for particel initalization improvement but not for updating particle movement. Still alot of work to be done here, should be clearer whenever you make a particle system what the params are and what they are used for. A big issue is performance, might need to manually make a pool of threads that iterate over a selection of the overall list of particles, the problem is how this should be done effectively