Redone project files, again, working
I just want to use someones elses code file, and I'm to imcompetent to do it, have spent days fiddling with project settings and files and executables and installing programs. Why does C++ have to be harder when it ccomes to handling files as well. This issue should be done in like 5 min, and I have spent 3 days and I just can't do it. I have no one to talk to, which is why I am talking to myself here, I feel incompetent and idiotic and embarrassed.