Personal made scene serialization
Altough I tried to implement this, even just a very simple one, it is so much work which would require alot of complexity, that is hard to implement cause my lack of experience with C++, and not worth my time. Can't prove it can be done sure, but would really just have wanted to used someone elses yaml serialization, and if I couldn't then just have moved on, but I have spent like a week on this topic and have come out with alot of nothing, feels bad. Will abandon serialization for now and probably for good for this disertation, time is running out, and there is so much else that would be nuch nicer to work with and give value, altough serialization would give the greatess of values and open a whole new lvl og development technique, but yeah gonna leave that for another time.