Fps counter updated
Updated the fps counter, twice, the first was to not clear the list of frames and push new frames into the list but create a list with a good amount of range and only assign and read the required range. Then I saw you could completly remove the need for a list. So better cleaner code and better performance. Also fooled around with headers, no fun, big headache and not managable at this point, headers are a labyrinth, also found that the pch is not in play anymore as that setting was probably not being pushed to the branch and when I had to get a fresh build it didn't have the pch settings. Have build so long without it and have no idea what has been added that is making it not compile with the pch anymore. Compiling the GameObject class takes like 16 seconds, really unacceptable compile times, but no idea how to resolve it, headers need a big redo and makeover.