Particle update
Found that I had hidden MyParticle classes that I was using instead of the visible ones, really odd having to use a self made class you can no longer see in the hiearchy in the solution explorer. Any way made some timing testing of the particle system and stuff. One note, the old without async particle effekt behaves different than when using it with async, the one with async looks better. Not sure what is going on, the color seems to be very faded without the async.