Physics jittery fix
Fixed the jittery of the physics system, now it actually works as intended, just had to really look through the physics code to notice duplications and slight errors, though I can understand why I thought there was a jittery issue cause of infrequent physics update calls, but no the delta time should and now does fix that completly. The physics is now free om jittery and inaccuracies and can finally be merged with the main branch. So happy, but yeah I knew how it should be done but I messed it up and tought something was wrong which was outside of my control, but no I just screwed up the code and confused myself, my bad