Trials of making a game
Well you can make a game, but you have to play within the limitations of the engine, and by that also the way the engine will fail to run.
Very time consuming creating even just a simple game, you're placing objects in the scene without being able to see it before it runs.
Errors, alot of code, it is just not a smooth process, and alot of things still have to be manually handled.
Very informative experience though, trying to make a game, seeing the difficulties and the hard facts, actually a big mistake on my part.
Of course you should try to take the perspective of the user, which to be fair is me, but dude you gotta try you're product, how else would you know if it's any good?